Create flask blueprints for ws/http (#100)
* Gather http routes in an exclusive module

* Gather events into a separate module
