+++ cast/app.py
@@ -1,36 +1,15 @@
 #!/usr/bin/env python3
-import argparse
-
-from flask import (
-    Flask,
-    render_template,
-    request,
-    send_from_directory,
-    redirect,
-    jsonify,
-)
-
-import sys
-
-import flask_socketio
-
-from flask_jwt_extended import (
-    JWTManager,
-    jwt_required,
-    create_access_token,
-    set_access_cookies,
-)
-
-from werkzeug.exceptions import BadRequest
-import json
-from .logger import Logging
-import pty
 import os
-import subprocess
+import sys
 import select
 import shlex
 import random
 import string
+import argparse
+
+from flask import Flask
+import flask_socketio
+from flask_jwt_extended import JWTManager
 
 __version__ = "0.0.1"
 
@@ -55,6 +34,10 @@ app.config["JWT_SECRET_KEY"] = random_string()
 
 jwt = JWTManager(app)
 socketio = flask_socketio.SocketIO(app)
+
+from .routes import http
+
+app.register_blueprint(http, url_prefix=r"")
 
 
 def read_and_forward_pty_output(session_id):
@@ -86,171 +69,6 @@ def read_and_forward_pty_output(session_id):
                     except OSError:
                         socketio.emit("disconnect", namespace="/cast")
                         sys.exit(0)
-
-
-@jwt.unauthorized_loader
-def unauthorized_response(callback):
-    return redirect("/")
-
-
-@jwt.invalid_token_loader
-def invalid_token(callback):
-    return redirect("/")
-
-
-@app.errorhandler(404)
-def page_not_found(error):
-    return render_template("404.html"), 404
-
-
-@app.route("/cast")
-@jwt_required
-def index():
-    log = Logging(app.config["current_session"])
-    log.make_log_folder()
-    return render_template("index.html")
-
-
-@app.route("/", methods=["GET", "POST"])
-def login():
-    if request.method == "GET":
-        return render_template("login.html")
-
-    elif request.method == "POST":
-        data = request.json
-        if data["password"] == app.config["passwd"]:
-            access_token = create_access_token(identity=os.getenv("JOB_ID"))
-            resp = jsonify({"login": True})
-            set_access_cookies(resp, access_token)
-            return resp
-        else:
-            return json.dumps(request.get_json()), 401
-
-
-@socketio.on("new-session", namespace="/cast")
-def new_session(data=None):
-    """To register session on WebSocket server
-    Similar to 'connect()', used for adding a new tab session
-    """
-    session_id = ""
-    if data is not None:
-        session_id = data["session_id"]
-    print("new-session: {}\n".format(session_id))
-
-    if session_id in app.config["sessions"]:
-        return
-
-    (child_pid, fd) = pty.fork()
-
-    if child_pid == 0:
-        # child: run system command
-        subprocess.run(app.config["cmd"])
-    else:
-        # parent: stream input/output from child to Socketio web channel
-        app.config["sessions"][session_id] = {}
-        app.config["sessions"][session_id]["fd"] = fd
-        app.config["sessions"][session_id]["child_pid"] = child_pid
-
-        cmd = " ".join(shlex.quote(c) for c in app.config["cmd"])
-
-        print("new-session: child pid is", child_pid)
-        print(
-            "new-session: starting background task with command `{}` to continously read "
-            "and forward pty output to client".format(cmd)
-        )
-        """
-        print(
-            f"new-session: starting background task with command `{cmd}` to continously read "
-            "and forward pty output to client"
-        )
-        """
-
-        socketio.start_background_task(
-            target=read_and_forward_pty_output, session_id=session_id
-        )
-
-        print("new-session: task started")
-
-
-@socketio.on("connect", namespace="/cast")
-def connect(data=None):
-    session_id = (
-        request.values.get("session_id")
-        if not request.values.get("session_id") is None
-        else ""
-    )
-    if session_id == "" and data is not None:
-        session_id = data["session_id"]
-        print("connect: {}\n".format(session_id))
-
-    # Create new session only when id not in records
-    if session_id in app.config["sessions"]:
-        return
-
-    (child_pid, fd) = pty.fork()
-
-    if child_pid == 0:
-        # child: start system command
-        subprocess.run(app.config["cmd"])
-    else:
-        # parent: print info, stream child input/output to socketio
-        # Store sessions by ssid
-        print("opening a new session")
-        app.config["sessions"] = {}
-        app.config["sessions"][session_id] = {}
-        app.config["sessions"][session_id]["fd"] = fd
-        app.config["sessions"][session_id]["child_pid"] = child_pid
-
-        cmd = " ".join(shlex.quote(c) for c in app.config["cmd"])
-
-        print("connect: child pid is", child_pid)
-        print(
-            "connect: starting background task with command `{}` to continously read "
-            "and forward pty output to client".format(cmd)
-        )
-        """
-        print(
-            f"new-session: starting background task with command `{cmd}` to continously read "
-            "and forward pty output to client"
-        )
-        """
-
-        # Output terminal message corresponding to ssid
-        socketio.start_background_task(
-            target=read_and_forward_pty_output, session_id=session_id
-        )
-
-        print("connect: task started")
-
-
-@socketio.on("client-input", namespace="/cast")
-def client_input(data):
-    # Update current session
-    app.config["current_session"] = data["session_id"]
-    print("input: {}".format(app.config["sessions"]))
-    log = Logging(app.config["current_session"])
-
-    if data["session_id"] in app.config["sessions"]:
-        file_desc = app.config["sessions"][data["session_id"]]["fd"]
-
-        if file_desc:
-            if data["input"] == "":
-                # When switching sessions, send a key to update terminal content
-                os.write(file_desc, b"\x00")
-            else:
-                log.write_log(data["input"])
-                os.write(file_desc, data["input"].encode())
-
-
-# This is the route handler for DOWNLOADING the log file. Maybe a bit buggy. Please report if found
-@app.route("/download/<string:file_path>")
-def download(file_path):
-    try:
-        return send_from_directory(
-            app.config["log_file"], filename=file_path, as_attachment=True
-        )
-    except BadRequest:
-        return 404
 
 
 def create_parser():
